// src/app/store/integration/Interface.ts
export interface TwoFactorIntegration {
  type: number | string;
  is_enable: number;
}
export interface removeIntegrationType {
  type: number | string;
}
export interface SlackNotificationType {
  types_of_slack_notification?: string;
  project_id?: number | string;
}
export interface IntegrationType {
  type: number;
}
export interface SLackIntegrationType {
  type: number;
  project_id: string;
}
/**
 * The type definition for the initial state of the auth slice.
 */
export type initialStateProps = {
  status: string;
  // supportStatus?: string;
  // details: any;
  // fetchStatus?: string;
  // list: SlackType[];
  // data?: any;
  // total_records: number;
  // subscription?: {};
  list?: {};
  // total_items: number;
  // assignedAgentDetail: any[];
  // assignAccManagerDetail?: any[];
  // supportlist?: any[];
  // Support_total_records?: number;
  // Support_total_items?: number;
};